pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rheed-realtime-analyzer"
version = "0.1.0"
description = "Sidecar analytics framework for real-time RHEED monitoring"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["RHEED", "PLD", "real-time", "analysis", "materials"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "build>=1.2",
  "twine>=5.0",
  "notebook>=7.0",
  "ipykernel>=6.0",
]

[project.scripts]
rheed-sidecar = "rheed_core.main:main"

[tool.setuptools.packages.find]
include = ["rheed_core*"]

[tool.setuptools.package-data]
rheed_core = ["config.yaml"]
